Fix copy error and CI error

+++ backend/bigg_database/management/commands/fix_pymysql.py
@@ -4,21 +4,25 @@ import shutil
 
 from django.core.management.base import BaseCommand, CommandError
 
-from backend.settings import BASE_DIR
-
 
 def main():
+    venv_dir = os.environ.get('VIRTUAL_ENV', None)
+    if not venv_dir:
+        raise CommandError('Can not determine venv dir')
     available_mysql_dir = os.path.join(
-        os.path.dirname(BASE_DIR), 'venv', 'lib', 'python3.*', 'site-packages', 'django', 'db', 'backends', 'mysql')
+        venv_dir, 'lib', 'python3.*', 'site-packages', 'django', 'db', 'backends', 'mysql')
     found_mysql_dir = glob.glob(available_mysql_dir)
     if not found_mysql_dir:
-        raise CommandError('Can not find the dir {}'.format(available_mysql_dir))
+        raise CommandError('Can not find the django mysql dir: {}'.format(available_mysql_dir))
     mysql_dir = found_mysql_dir[0]
     base_path = os.path.join(mysql_dir, 'base.py')
     operation_path = os.path.join(mysql_dir, 'operations.py')
 
-    shutil.copyfile(base_path, base_path + '.bak')
-    shutil.copyfile(operation_path, operation_path + '.bak')
+    if not os.path.isfile(base_path + '.bak'):
+        shutil.copyfile(base_path, base_path + '.bak')
+
+    if not os.path.isfile(operation_path + '.bak'):
+        shutil.copyfile(operation_path, operation_path + '.bak')
 
     content = []
     with open(base_path, 'r') as file:
@@ -32,6 +36,8 @@ def main():
                 content.append('    pass\n')
     with open(base_path, 'w') as file:
         file.writelines(content)
+
+    content = []
     with open(operation_path, 'r') as file:
         for line in file:
             if line.strip() != r"query = query.decode(errors='replace')":
